refactor into a single "init" function
previously, "store" was a separate method which would be called
on-demand, using the player's current position to ensure the position
was loaded before placing a command-block.

The new "if loaded" test can be used to ensure commands aren't run until
the chunks are fully loaded, which allows us to place the command blocks
in the spawn chunks.

As part of this change, only change the commandBlockOutput /
sendCommandFeedback settings when needed, ensuring we don't override
existing settings.